requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
python-dotenv
